refactor: wallet name screens
