+++ packages/core/__tests__/screens/RenameWallet.test.tsx
@@ -2,10 +2,10 @@ import { render } from '@testing-library/react-native'
 import React from 'react'
 
 import { StoreProvider } from '../../src'
-import NameWallet from '../../src/screens/NameWallet'
 import { testIdWithKey } from '../../src/utils/testable'
 import { testDefaultState } from '../contexts/store'
 import { BasicAppContext } from '../helpers/app'
+import RenameWallet from '../../src/screens/RenameWallet'
 
 jest.mock('react-native-permissions', () => require('react-native-permissions/mock'))
 jest.mock('react-native-vision-camera', () => {
@@ -19,22 +19,21 @@ jest.mock('react-native-device-info', () => {
   }
 })
 
-describe('NameWallet Screen', () => {
-  test('LimitedInput and continue button are present', async () => {
+describe('RenameWallet Screen', () => {
+  test('Check non-onboarding rendering', async () => {
     const tree = render(
       <StoreProvider initialState={testDefaultState}>
         <BasicAppContext>
-          <NameWallet />
+          <RenameWallet/>
         </BasicAppContext>
       </StoreProvider>
     )
 
-    const WalletNameInput = await tree.getByTestId(testIdWithKey('NameInput'))
-    const ContinueButton = await tree.getByTestId(testIdWithKey('Continue'))
+    const SaveButton = await tree.getByTestId(testIdWithKey('Save'))
+    const CancelButton = await tree.getByTestId(testIdWithKey('Cancel'))
 
-    expect(WalletNameInput).not.toBeNull()
-    expect(ContinueButton).not.toBeNull()
+    expect(SaveButton).not.toBeNull()
+    expect(CancelButton).not.toBeNull()
     expect(tree).toMatchSnapshot()
-
   })
 })
